+++ gazebo/gui/ModelMaker_TEST.hh
@@ -42,6 +42,14 @@ class ModelMaker_TEST : public QTestFixture
 
   /// \brief Test creating a nested model by copying another nested model.
   private slots: void FromNestedModel();
+
+  /// \brief Test creating a model with spaces in the file name and entity
+  /// names.
+  private slots: void FromModelWithSpaces();
+
+  /// \brief Test creating a nested model with frame semantics (SDFormat 1.7)
+  /// from a file
+  private slots: void FromNestedModelFileWithFrameSemantics();
 };
 
 #endif