+++ gazebo/rendering/Material.cc
@@ -14,6 +14,8 @@
  * limitations under the License.
  *
 */
+#include <ignition/common/Profiler.hh>
+
 #include "gazebo/rendering/ogre_gazebo.h"
 #include "gazebo/common/Console.hh"
 #include "gazebo/rendering/RenderEngine.hh"
@@ -139,6 +141,7 @@ void Material::CreateMaterials()
 //////////////////////////////////////////////////
 void Material::Update(const gazebo::common::Material *_mat)
 {
+  IGN_PROFILE("rendering::Material::Update");
   Ogre::MaterialPtr matPtr;
 
   if (Ogre::MaterialManager::getSingleton().resourceExists(_mat->GetName()))