+++ gazebo/rendering/UserCamera.cc
@@ -15,6 +15,7 @@
  *
 */
 #include <boost/bind.hpp>
+#include <ignition/common/Profiler.hh>
 #include <ignition/math/Color.hh>
 #include <ignition/math/Vector2.hh>
 
@@ -230,6 +231,7 @@ void UserCamera::SetWorldPose(const ignition::math::Pose3d &_pose)
 //////////////////////////////////////////////////
 void UserCamera::Update()
 {
+  IGN_PROFILE("rendering::UserCamera::Update");
   Camera::Update();
 
   if (this->dataPtr->viewController)
@@ -248,6 +250,7 @@ void UserCamera::AnimationComplete()
 //////////////////////////////////////////////////
 void UserCamera::Render(const bool /*_force*/)
 {
+  IGN_PROFILE("rendering::UserCamera::Render");
   if (this->initialized)
   {
     this->newData = true;
@@ -258,6 +261,7 @@ void UserCamera::Render(const bool /*_force*/)
 //////////////////////////////////////////////////
 void UserCamera::PostRender()
 {
+  IGN_PROFILE("rendering::UserCamera::PostRender");
   Camera::PostRender();
 }