+++ gazebo/sensors/GpuRaySensor.cc
@@ -673,6 +673,7 @@ void GpuRaySensor::PrerenderEnded()
 //////////////////////////////////////////////////
 void GpuRaySensor::Render()
 {
+  IGN_PROFILE("sensors::GpuRaySensor::Render");
   if (this->useStrictRate)
   {
     if (!this->dataPtr->renderNeeded)